[DaoPTA][medicine] module colaps with Cao

+++ src/main/java/com/example/retro_care/kind_of_medicine/repository/IKindOfMedicineRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IKindOfMedicineRepository extends JpaRepository<KindOfMedicine,Integer> {
+public interface IKindOfMedicineRepository extends JpaRepository<KindOfMedicine,Long> {
 //    @Query(value = "select code,name from kind_of_medicine where code like :searchCode && name like :searchName")
 }
